uv: runner: force reinstall ingestr on run
This is necessary to pull latest changes when working with a
local mirror of ingestr. The penalty in case of non-local use
is marginal, since uv caches previous installs and verifies
quickly (<30ms) that it doesn't need to do anything.

The alternative to this is to do hacky and sensitive manipulation
of site-package directory created by uv for it's tools.
(and quite possibly in other metadata files managed by uv).
This approach would also be vulnerable to breaking in the future
so it's decidedly inferior for what we're trying to achieve.