lint: refactor asset linting.
When we run
```
bruin validate ./pipeline
```

Asset rules are invoked via rule.Validate (type PipelineValidator).

This was causing selector logic in policies to fail, since these
were implemented as decorators on top of different validator
interfaces.

This commit changes lint.LintPipeline to use the rule.ValidateAsset
when validation assets. This change eliminates the need for wrapping
asset level validators with `CallFuncForEveryAsset`, as well as
promoting rule.ApplicableLevels to a first class citizen of the
lint package.